Move DeviceTrait implementation to wasapi::device
